Ticket 47617 - replication changelog trimming setting validation
Bug Description:  The changelog trimming interval setting is not validated.
                  Multiple and invalid values were allowed.  The new interval
                  was also not taking effect until the old timeout completed.

Fix Description:  Add the config attribute to the core schema to addreess
                  the multivalue issue.  Added the setting validation checks,
                  and notify the trimming thread when the interval is changed.

                  Also applied these validatations to the changelog compact interval.

https://fedorahosted.org/389/ticket/47617

Reviewed by: nhosoi(Thanks!)

+++ ldap/servers/plugins/replication/cl5.h
@@ -56,8 +56,8 @@ typedef struct changelog5Config
 /* the changelog DB configuration parameters are defined as CL5DBConfig in cl5_api.h */
 	CL5DBConfig dbconfig;	
 	char *symmetricKey;
-	int compactInterval;
-	int trimInterval;
+	long compactInterval;
+	long trimInterval;
 }changelog5Config;
 
 /* initializes changelog*/
